bot/requirements.txt
aiohttp
aiortc
pipecatcloud
pipecat-ai[cartesia,daily,deepgram,google,local-smart-turn-v3,openai,silero,webrtc]>=0.0.89
pipecat-ai-small-webrtc-prebuilt
pipecat-ai-tail
python-dotenv~=1.0.1
opencv-python>=4.8.0
numpy>=1.24.0
